+++ quint/src/parsing/ToIrListener.ts
@@ -38,9 +38,10 @@ import { TerminalNode } from 'antlr4ts/tree/TerminalNode'
 import { QuintTypeDef } from '../ir/quintIr'
 import { zip } from '../util'
 import { QuintError } from '../quintError'
-import { lowercaseTypeError, tooManySpreadsError, undeclaredTypeParamsError } from './parseErrors'
+import { lowercaseTypeError, mapSyntaxError, tooManySpreadsError, undeclaredTypeParamsError } from './parseErrors'
 import { Loc } from '../ErrorMessage'
 import assert, { fail } from 'assert'
+import { typeToString } from '../ir/IRprinting'
 
 /**
  * An ANTLR4 listener that constructs QuintIr objects out of the abstract
@@ -1045,8 +1046,23 @@ export class ToIrListener implements QuintListener {
     // The next type on the stack after the args should be the applied
     // type constructor
     const ctor: QuintConstType = { id: this.getId(ctx), kind: 'const', name: ctx._typeCtor.text }
-    const typeApp: QuintAppType = { id, kind: 'app', ctor, args }
-    this.typeStack.push(typeApp)
+
+    // Check for Map[a, b] syntax
+    if (ctor.name === 'Map' && args.length === 2) {
+      // Extract the key and value type names
+      const keyType = typeToString(args[0])
+      const valueType = typeToString(args[1])
+
+      // Add error with fix suggestion
+      this.errors.push(mapSyntaxError(id, keyType, valueType))
+
+      // Create a function type instead (which is the correct representation for maps)
+      const funType: QuintType = { id, kind: 'fun', arg: args[0], res: args[1] }
+      this.typeStack.push(funType)
+    } else {
+      const typeApp: QuintAppType = { id, kind: 'app', ctor, args }
+      this.typeStack.push(typeApp)
+    }
   }
 
   // TODO: replace with general type application